refactor: create useHistoryPanelLogic
